feat: Refactor food recognition upload functionality into a modal
- Removed the upload component and migrated its logic to a new UploadModal component.
- Integrated upload modal with barcode, QR, and OCR scanners.
- Added context management for upload modal visibility.
- Updated bottom navigation and discover button to trigger the upload modal.
- Enhanced user experience with haptic feedback and animations.
